+++ java-client/src/main/java/io/sphere/client/exceptions/DuplicateSkuException.java
@@ -1,0 +1,7 @@
+package io.sphere.client.exceptions;
+
+public class DuplicateSkuException  extends SphereException {
+    public DuplicateSkuException(String sku) {
+        super("SKU already in use: " + sku);
+    }
+}