+++ java-client/src/main/java/io/sphere/client/shop/model/InventoryEntry.java
@@ -1,13 +1,13 @@
 package io.sphere.client.shop.model;
 
 import com.google.common.base.Optional;
-import io.sphere.client.model.EmptyReference;
-import io.sphere.client.model.Reference;
 import io.sphere.client.model.VersionedId;
+import io.sphere.internal.util.Iso8601JsonSerializer;
 import net.jcip.annotations.Immutable;
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 import org.codehaus.jackson.annotate.JsonProperty;
-
+import org.codehaus.jackson.map.annotate.JsonSerialize;
+import org.joda.time.DateTime;
 import javax.annotation.Nonnull;
 
 /** Inventory information for a specific product variant in a catalog. */
@@ -16,13 +16,12 @@ import javax.annotation.Nonnull;
 public class InventoryEntry {
     @Nonnull private String id;
     @JsonProperty("version") private int version;
-    @Nonnull private String productId;
-    private int variantId;
-    private int quantityOnStock;
-    private int availableQuantity;
-    @Nonnull private Reference<Catalog> catalog = EmptyReference.create("catalog");
     private String sku = "";
+    private long quantityOnStock;
+    private long availableQuantity;
     private Integer restockableInDays;
+    @JsonSerialize(using = Iso8601JsonSerializer.class)
+    private DateTime expectedDelivery;
 
     // for JSON deserializer
     protected InventoryEntry() {}
@@ -33,21 +32,11 @@ public class InventoryEntry {
     /** The {@link #getId() id} plus version. */
     @Nonnull public VersionedId getIdAndVersion() { return VersionedId.create(id, version); }
 
-    /** The product id to which the inventory entry belongs. */
-    @Nonnull public String getProductId() { return productId; }
-
-    /** The product variant id to which the inventory entry belongs. */
-    public int getVariantId() { return variantId; }
-
     /** Current quantity on stock. */
-    public int getQuantityOnStock() { return quantityOnStock; }
+    public long getQuantityOnStock() { return quantityOnStock; }
 
     /** Current available quantity (quantity on stock minus the quantity of reserved items). */
-    public int getAvailableQuantity() { return availableQuantity; }
-
-    /** The catalog to which the inventory entry belongs.
-     * @return The catalog of null in case of the master catalog. */
-    @Nonnull public Reference<Catalog> getCatalog() { return catalog; }
+    public long getAvailableQuantity() { return availableQuantity; }
 
     /** The SKU of the product to which the inventory entry belongs. */
     public String getSku() { return sku; }
@@ -56,18 +45,20 @@ public class InventoryEntry {
      *  @return The number of days or {@link Optional#absent()} if not specified. */
     @Nonnull public Optional<Integer> getRestockableInDays() { return Optional.fromNullable(restockableInDays); }
 
+    public DateTime getExpectedDelivery() {
+        return expectedDelivery;
+    }
+
     @Override
     public String toString() {
         return "InventoryEntry{" +
                 "id='" + id + '\'' +
                 ", version=" + version +
-                ", productId='" + productId + '\'' +
-                ", variantId=" + variantId +
+                ", sku='" + sku + '\'' +
                 ", quantityOnStock=" + quantityOnStock +
                 ", availableQuantity=" + availableQuantity +
-                ", catalog=" + catalog +
-                ", sku='" + sku + '\'' +
                 ", restockableInDays=" + restockableInDays +
+                ", expectedDelivery=" + expectedDelivery +
                 '}';
     }
 }