+++ java-client/src/main/java/io/sphere/internal/CustomObjectServiceImpl.java
@@ -5,6 +5,7 @@ import io.sphere.client.CommandRequest;
 import io.sphere.client.DeleteRequest;
 import io.sphere.client.FetchRequest;
 import io.sphere.client.ProjectEndpoints;
+import io.sphere.client.model.QueryResult;
 import io.sphere.client.shop.CustomObjectService;
 import io.sphere.client.model.CustomObject;
 import io.sphere.client.shop.ApiMode;
@@ -12,13 +13,9 @@ import io.sphere.internal.request.RequestFactory;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.codehaus.jackson.type.TypeReference;
 
-public class CustomObjectServiceImpl extends ProjectScopedAPI implements CustomObjectService {
-    private final RequestFactory requestFactory;
-    private final TypeReference<CustomObject> typeReference = new TypeReference<CustomObject>() {};
-
+public class CustomObjectServiceImpl extends ProjectScopedAPI<CustomObject> implements CustomObjectService {
     public CustomObjectServiceImpl(RequestFactory requestFactory, ProjectEndpoints endpoints) {
-        super(endpoints);
-        this.requestFactory = requestFactory;
+        super(requestFactory, endpoints, new TypeReference<CustomObject>() {}, new TypeReference<QueryResult<CustomObject>>() { });
     }
 
     @Override