+++ java-client/src/main/java/io/sphere/internal/ReviewServiceImpl.java
@@ -10,19 +10,15 @@ import io.sphere.client.shop.ApiMode;
 import io.sphere.client.shop.ReviewService;
 import io.sphere.client.shop.model.Review;
 import io.sphere.client.shop.model.ReviewUpdate;
-import io.sphere.internal.command.Command;
 import io.sphere.internal.command.ReviewCommands;
 import io.sphere.internal.command.UpdateCommand;
 import io.sphere.internal.request.RequestFactory;
 import com.google.common.base.Optional;
 import org.codehaus.jackson.type.TypeReference;
 
-public class ReviewServiceImpl extends ProjectScopedAPI implements ReviewService {
-    private final RequestFactory requestFactory;
-
+public class ReviewServiceImpl extends ProjectScopedAPI<Review> implements ReviewService {
     public ReviewServiceImpl(RequestFactory requestFactory, ProjectEndpoints endpoints) {
-        super(endpoints);
-        this.requestFactory = requestFactory;
+        super(requestFactory, endpoints, new TypeReference<Review>() {}, new TypeReference<QueryResult<Review>>() { });
     }
 
     @Override public FetchRequest<Review> byId(String id) {
@@ -38,10 +34,7 @@ public class ReviewServiceImpl extends ProjectScopedAPI implements ReviewService
     }
 
     @Override public QueryRequest<Review> query() {
-        return requestFactory.createQueryRequest(
-                endpoints.reviews.root(),
-                Optional.<ApiMode>absent(),
-                new TypeReference<QueryResult<Review>>() {});
+        return queryImpl(endpoints.reviews.root());
     }
 
     @Override public QueryRequest<Review> forCustomer(String customerId) {
@@ -76,10 +69,5 @@ public class ReviewServiceImpl extends ProjectScopedAPI implements ReviewService
         return createCommandRequest(
                 endpoints.reviews.byId(reviewId.getId()),
                 new UpdateCommand<ReviewCommands.ReviewUpdateAction>(reviewId.getVersion(), update));
-    }
-
-    /** Helper to save some repetitive code in this class. */
-    private CommandRequest<Review> createCommandRequest(String url, Command command) {
-        return requestFactory.createCommandRequest(url, command, new TypeReference<Review>() {});
     }
 }