+++ java-client/src/main/java/io/sphere/internal/ShippingMethodServiceImpl.java
@@ -14,13 +14,11 @@ import io.sphere.internal.request.RequestFactory;
 import com.google.common.base.Optional;
 import org.codehaus.jackson.type.TypeReference;
 
-public class ShippingMethodServiceImpl extends ProjectScopedAPI implements ShippingMethodService {
-    private final RequestFactory requestFactory;
-    private final static String expandZonesPath = "zoneRates[*].zone"; 
+public class ShippingMethodServiceImpl extends ProjectScopedAPI<ShippingMethod> implements ShippingMethodService {
+    private final static String expandZonesPath = "zoneRates[*].zone";
 
     public ShippingMethodServiceImpl(RequestFactory requestFactory, ProjectEndpoints endpoints) {
-        super(endpoints);
-        this.requestFactory = requestFactory;
+        super(requestFactory, endpoints, new TypeReference<ShippingMethod>() {}, new TypeReference<QueryResult<ShippingMethod>>() {});
     }
 
     @Override public FetchRequest<ShippingMethod> byId(String id) {
@@ -36,10 +34,7 @@ public class ShippingMethodServiceImpl extends ProjectScopedAPI implements Shipp
     }
 
     @Override public QueryRequest<ShippingMethod> query() {
-        return requestFactory.createQueryRequest(
-                endpoints.shippingMethods.root(),
-                Optional.<ApiMode>absent(),
-                new TypeReference<QueryResult<ShippingMethod>>() {}).expand(expandZonesPath);
+        return queryImpl(endpoints.shippingMethods.root());
     }
 
     @Override