+++ java-client/src/main/java/io/sphere/internal/request/CommandRequestImpl.java
@@ -6,13 +6,16 @@ import io.sphere.client.exceptions.SphereBackendException;
 import io.sphere.client.exceptions.SphereException;
 import io.sphere.client.SphereResult;
 import io.sphere.internal.command.Command;
+import io.sphere.internal.util.Iso8601JsonSerializer;
 import io.sphere.internal.util.Util;
 import io.sphere.client.CommandRequest;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import net.jcip.annotations.Immutable;
+import org.codehaus.jackson.Version;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.codehaus.jackson.map.ObjectWriter;
+import org.codehaus.jackson.map.module.SimpleModule;
 import org.codehaus.jackson.type.TypeReference;
 
 import javax.annotation.Nonnull;
@@ -42,7 +45,12 @@ public class CommandRequestImpl<T> implements CommandRequest<T>, TestableRequest
         if (requestHolder == null) throw new NullPointerException("requestHolder");
         if (command == null) throw new NullPointerException("command");
         if (jsonParserTypeRef == null) throw new NullPointerException("jsonParserTypeRef");
-        ObjectWriter jsonWriter = new ObjectMapper().writer();
+        //TODO the construction of an ObjectMapper and a Module per Request is expensive
+        ObjectMapper mapper = new ObjectMapper();
+        SimpleModule testModule = new SimpleModule("Iso8601JsonSerializerModule", new Version(1, 0, 0, null));
+        testModule.addSerializer(new Iso8601JsonSerializer());
+        mapper.registerModule(testModule);
+        ObjectWriter jsonWriter = mapper.writer();
         try {
             this.requestHolder = requestHolder.setBody(jsonWriter.writeValueAsString(command));
         } catch (IOException e) {