+++ java-client/src/main/java/io/sphere/internal/util/Iso8601JsonSerializer.java
@@ -1,0 +1,24 @@
+package io.sphere.internal.util;
+
+import org.codehaus.jackson.JsonGenerator;
+import org.codehaus.jackson.JsonProcessingException;
+import org.codehaus.jackson.map.SerializerProvider;
+import org.codehaus.jackson.map.ser.std.SerializerBase;
+import org.joda.time.DateTime;
+import org.joda.time.format.DateTimeFormatter;
+import org.joda.time.format.ISODateTimeFormat;
+
+import java.io.IOException;
+
+public class Iso8601JsonSerializer extends SerializerBase<DateTime> {
+    private static DateTimeFormatter FORMATTER = ISODateTimeFormat.dateTime();
+
+    public Iso8601JsonSerializer() {
+        super(DateTime.class);
+    }
+
+    @Override
+    public void serialize(DateTime value, JsonGenerator gen, SerializerProvider arg2) throws IOException, JsonProcessingException {
+        gen.writeString(FORMATTER.print(value));
+    }
+}